Add get method to the involvement API
