+++ src/modules/display.js
@@ -2,7 +2,7 @@ import getShows from './getShows.js';
 import thumbsUp from '../images/thumbsup.svg';
 import commentImg from '../images/comment.svg';
 import displayModal from './displayModal.js';
-import addNewLike from './involvementAPI.js';
+import addNewLike from './likes.js';
 
 const displayScreen = async () => {
   const mostRatedShows = await getShows();
