--- src/modules/involvementAPI.js
@@ -1,17 +1,0 @@
-const addNewLike = async (id) => {
-  await fetch(`https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/wTuKRCdWCi0TS2pluTow/likes?item_id=${id}`,
-    {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        item_id: id,
-      }),
-    });
-//     .then((response) => {
-//     console.log(response.json());
-//   });
-};
-
-export default addNewLike;
